refactor: move all ProviderModules to DataModule
