+++ core/data/src/main/java/com/projectlab/core/data/di/DataModule.kt
@@ -5,16 +5,22 @@ import com.google.firebase.firestore.FirebaseFirestore
 import com.google.firebase.firestore.ktx.firestore
 import com.google.firebase.ktx.Firebase
 import com.projectlab.core.data.Constants.REFERENCE_USERS
+import com.projectlab.core.data.repository.AmadeusTokenProviderImpl
 import com.projectlab.core.data.repository.FirestoreFlightRepositoryImpl
 import com.projectlab.core.data.repository.FirestoreFlightSegmentRepositoryImpl
 import com.projectlab.core.data.repository.FirestoreHotelRepositoryImpl
 import com.projectlab.core.data.repository.FirestoreItineraryRepositoryImpl
 import com.projectlab.core.data.repository.FirestoreUserRepositoryImpl
+import com.projectlab.core.data.repository.OnboardingFlagProviderImpl
+import com.projectlab.core.data.repository.SearchHistoryProviderImpl
 import com.projectlab.core.data.repository.UsersRepositoryImpl
 import com.projectlab.core.domain.repository.FlightRepository
 import com.projectlab.core.domain.repository.FlightSegmentRepository
 import com.projectlab.core.domain.repository.HotelRepository
 import com.projectlab.core.domain.repository.ItineraryRepository
+import com.projectlab.core.domain.repository.OnboardingFlagProvider
+import com.projectlab.core.domain.repository.SearchHistoryProvider
+import com.projectlab.core.domain.repository.TokenProvider
 import com.projectlab.core.domain.repository.UserRepository
 import com.projectlab.core.domain.repository.UsersRepository
 import dagger.Binds
@@ -36,6 +42,23 @@ import javax.inject.Singleton
 @InstallIn(SingletonComponent::class)
 @Module
 abstract class DataModule {
+    @Binds
+    @Singleton
+    abstract fun bindOnboardingFlagProvider(
+        onboardingFlagProviderImpl: OnboardingFlagProviderImpl,
+    ): OnboardingFlagProvider
+
+    @Binds
+    @Singleton
+    abstract fun bindAmadeusTokenProvider(
+        amadeusTokenProviderImpl: AmadeusTokenProviderImpl,
+    ): TokenProvider
+
+    @Binds
+    @Singleton
+    abstract fun bindSearchHistoryProvider(
+        searchHistoryProviderImpl: SearchHistoryProviderImpl,
+    ): SearchHistoryProvider
 
     @Binds
     abstract fun bindUserRepository(
